:rotating_light: Fix `ty` and `ruff` issues (#199)
* :art: Run Ruff check and format on the whole source

* :label: Fix all ty(invalid-assignment) issues

* :label: Fix or supress ty issues

* :bug: Fix unresolved dict() calls in some image2text models

* :bug: Fix preprocessor bug in `Model`

* :bug: Fix f1 zero division compatibility bug

* :bug: Fix whisper config to_dict bug

* :rotating_light: Ruff the source code

+++ hezar/models/mask_filling/roberta/roberta_mask_filling_config.py
@@ -15,7 +15,7 @@ class RobertaMaskFillingConfig(ModelConfig):
     hidden_act: str = "gelu"
     hidden_dropout_prob: float = 0.1
     hidden_size: int = 768
-    initializer_range: int = 0.02
+    initializer_range: float = 0.02
     intermediate_size: int = 3072
     layer_norm_eps: float = 1e-12
     max_position_embeddings: int = 514
